refactor(#86): member성공,예외 코드 변경
