+++ src/main/java/friendy/community/domain/member/controller/code/MemberSuccessCode.java
@@ -8,9 +8,9 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum MemberSuccessCode implements ApiCode {
 
-    SIGN_UP_SUCCESS(1101, "회원가입에 성공했습니다."),
-    CHANGE_PASSWORD_SUCCESS(1102, "비밀번호 변경에 성공했습니다."),
-    GET_MEMBER_INFO_SUCCESS(1103, "프로필 조회에 성공했습니다.");
+    SIGN_UP_SUCCESS(1201, "회원가입에 성공했습니다."),
+    CHANGE_PASSWORD_SUCCESS(1202, "비밀번호 변경에 성공했습니다."),
+    GET_MEMBER_INFO_SUCCESS(1203, "프로필 조회에 성공했습니다.");
 
     private final int code;
     private final String message;
